+++ MultiVendor-BackEnd/database/migrations/2022_03_23_203732_create_orders_table.php
@@ -18,7 +18,7 @@ class CreateOrdersTable extends Migration
             $table->bigInteger('customer_id')->unsigned()->index();
             $table->foreign('customer_id')->references('id')->on('customers')->onDelete('cascade');
             $table->timestamps();
-            $table->bigInteger('total_price')->unsigned();
+            $table->float('total_price')->unsigned();
             $table->bigInteger('status')->unsigned();
 
         });
