+++ MultiVendor-BackEnd/database/migrations/2022_03_23_211509_create_balances_table.php
@@ -17,9 +17,9 @@ class CreateBalancesTable extends Migration
             $table->id();
             $table->bigInteger('vendor_id')->unsigned()->index();
             $table->foreign('vendor_id')->references('id')->on('vendors')->onDelete('cascade');
-            $table->bigInteger('total_sales')->unsigned()->default(0);
-            $table->bigInteger('received_ammount')->unsigned()->default(0);
-            $table->bigInteger('remaining_ammount')->unsigned()->default(0);
+            $table->float('total_sales')->unsigned()->default(0);
+            $table->float('received_ammount')->unsigned()->default(0);
+            $table->float('remaining_ammount')->unsigned()->default(0);
             $table->timestamps();
         });
     }
